+++ internal/stats/http.go
@@ -4,14 +4,24 @@ package stats
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
+	"strconv"
 	"time"
 
 	"github.com/AdguardTeam/AdGuardHome/internal/aghalg"
 	"github.com/AdguardTeam/AdGuardHome/internal/aghhttp"
 	"github.com/AdguardTeam/AdGuardHome/internal/aghnet"
 	"github.com/AdguardTeam/golibs/timeutil"
+	"github.com/AdguardTeam/golibs/validate"
 )
+
+// queryKeyRecent is the key of the query parameter that contains the lookback
+// interval for statistics.
+const queryKeyRecent = "recent"
+
+// millisecondsInHour contains number of milliseconds in one hour.
+const millisecondsInHour = int64(time.Hour / time.Millisecond)
 
 // topAddrs is an alias for the types of the TopFoo fields of statsResponse.
 // The key is either a client's address or a requested address.
@@ -53,16 +63,24 @@ func (s *StatsCtx) handleStats(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	l := s.logger
 
-	var (
-		resp *StatsResp
-		ok   bool
-	)
+	var limit time.Duration
 	func() {
 		s.confMu.RLock()
 		defer s.confMu.RUnlock()
 
-		resp, ok = s.getData(uint32(s.limit.Hours()))
+		limit = s.limit
 	}()
+
+	recent := r.URL.Query().Get(queryKeyRecent)
+
+	limit, err := parseRecent(recent, limit)
+	if err != nil {
+		aghhttp.ErrorAndLog(ctx, l, r, w, http.StatusBadRequest, "%s", err)
+
+		return
+	}
+
+	resp, ok := s.getData(uint32(limit.Hours()))
 
 	l.DebugContext(ctx, "prepared data", "elapsed", time.Since(start))
 
@@ -76,6 +94,31 @@ func (s *StatsCtx) handleStats(w http.ResponseWriter, r *http.Request) {
 	}
 
 	aghhttp.WriteJSONResponseOK(ctx, l, w, r, resp)
+}
+
+// parseRecent parses and validates the value of the recent URL parameter.  If
+// the parameter is empty, the original limit is returned.
+func parseRecent(recent string, limit time.Duration) (parsedLimit time.Duration, err error) {
+	if recent == "" {
+		return limit, nil
+	}
+
+	recentMs, err := strconv.ParseInt(recent, 10, 64)
+	if err != nil {
+		return 0, fmt.Errorf("%s: parsing interval: %s", queryKeyRecent, err)
+	}
+
+	err = validate.InRange(queryKeyRecent, recentMs, millisecondsInHour, limit.Milliseconds())
+	if err != nil {
+		// Don't wrap the error since it's already informative enough as is.
+		return 0, err
+	}
+
+	if recentMs%millisecondsInHour != 0 {
+		return 0, fmt.Errorf("%s: must be a multiple of 1 hour", queryKeyRecent)
+	}
+
+	return time.Duration(recentMs) * time.Millisecond, nil
 }
 
 // configResp is the response to the GET /control/stats_info.